--- a/tests/relational/test_artifacts.py
+++ b/tests/relational/test_artifacts.py
@@ -1,47 +1,78 @@
+import os
+import shutil
 import tarfile
 import tempfile
 from pathlib import Path
 from unittest.mock import Mock
 
-from gretel_trainer.relational.artifacts import ArtifactCollection, add_to_tar
+import pytest
+
+from gretel_trainer.relational.artifacts import (
+    ArtifactCollection,
+    archive_items,
+    archive_nested_dir,
+)
 
 
-def test_makes_new_archive():
-    with tempfile.TemporaryDirectory() as tmpdir, tempfile.NamedTemporaryFile() as tf1:
-        archive_path = Path(tmpdir) / "archive.tar.gz"
-        add_to_tar(archive_path, Path(tf1.name), "tf1")
-
-        with tarfile.open(archive_path, "r:gz") as tar:
-            assert len(tar.getnames()) == 1
+@pytest.fixture()
+def out_dir():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        yield Path(tmpdir)
 
 
-def test_appends_to_existing_archive():
-    with tempfile.TemporaryDirectory() as tmpdir, tempfile.NamedTemporaryFile() as tf1, tempfile.NamedTemporaryFile() as tf2:
-        archive_path = Path(tmpdir) / "archive.tar.gz"
-        add_to_tar(archive_path, Path(tf1.name), "tf1")
-        add_to_tar(archive_path, Path(tf2.name), "tf2")
-
-        with tarfile.open(archive_path, "r:gz") as tar:
-            assert len(tar.getnames()) == 2
+@pytest.fixture()
+def tempfiles():
+    with tempfile.NamedTemporaryFile() as tf1, tempfile.NamedTemporaryFile() as tf2:
+        yield [Path(tf1.name), Path(tf2.name)]
 
 
-def test_overwrites_existing_file_in_archive():
-    with tempfile.TemporaryDirectory() as tmpdir, tempfile.NamedTemporaryFile() as tf1:
-        archive_path = Path(tmpdir) / "archive.tar.gz"
-        add_to_tar(archive_path, Path(tf1.name), "tf1")
-        add_to_tar(archive_path, Path(tf1.name), "tf1")
+def test_archiving_items(out_dir, tempfiles):
+    tgz = out_dir / "out.tar.gz"
+    archive_items(tgz, tempfiles)
+    print(tgz.exists())
 
-        with tarfile.open(archive_path, "r:gz") as tar:
-            assert len(tar.getnames()) == 1
+    with tarfile.open(tgz, "r:gz") as tar:
+        assert set(tar.getnames()) == {
+            ".",
+            f"./{tempfiles[0].name}",
+            f"./{tempfiles[1].name}",
+        }
 
-    # Overwrite is based on provided arcname, not file contents
-    with tempfile.TemporaryDirectory() as tmpdir, tempfile.NamedTemporaryFile() as tf1, tempfile.NamedTemporaryFile() as tf2:
-        archive_path = Path(tmpdir) / "archive.tar.gz"
-        add_to_tar(archive_path, Path(tf1.name), "name")
-        add_to_tar(archive_path, Path(tf2.name), "name")
 
-        with tarfile.open(archive_path, "r:gz") as tar:
-            assert len(tar.getnames()) == 1
+def test_archive_nested_dir(out_dir, tempfiles):
+    tgz = out_dir / "out.tar.gz"
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        dir_to_archive = Path(tmpdir)
+        for tfile in tempfiles:
+            shutil.copy(tfile, dir_to_archive)
+
+        # When the provided targz path does not yet exist, creates a new archive with a nested archive
+        archive_nested_dir(tgz, dir_to_archive, "id1")
+
+        with tarfile.open(tgz, "r:gz") as tar:
+            assert set(tar.getnames()) == {".", "./id1.tar.gz"}
+
+            with tempfile.TemporaryDirectory() as td:
+                tar.extract("./id1.tar.gz", td)
+                with tarfile.open(f"{td}/id1.tar.gz", "r:gz") as nested:
+                    assert set(nested.getnames()) == {
+                        ".",
+                        f"./{tempfiles[0].name}",
+                        f"./{tempfiles[1].name}",
+                    }
+
+        # Later, new directories can be added to the existing archive
+        archive_nested_dir(tgz, dir_to_archive, "id2")
+
+        with tarfile.open(tgz, "r:gz") as tar:
+            assert set(tar.getnames()) == {".", "./id1.tar.gz", "./id2.tar.gz"}
+
+        # In the case of a name conflict, the added dir overwrites/replaces an existing dir
+        archive_nested_dir(tgz, dir_to_archive, "id2")
+
+        with tarfile.open(tgz, "r:gz") as tar:
+            assert set(tar.getnames()) == {".", "./id1.tar.gz", "./id2.tar.gz"}
 
 
 def test_uploads_path_to_project_and_stores_artifact_key():